requirements.txt
httpx
python-dotenv
python-telegram-bot==20.8
fastapi
uvicorn
